requirements.txt
Flask
Flask-Cors
ultralytics
torch
torchvision
easyocr
Pillow
numpy
pyngrok
lap